Resolve Error.

+++ src/Solution01/Solution01.java
@@ -8,7 +8,7 @@ public class Solution01 {
         double n;
         double sum = 0.0;
         Scanner scan = new Scanner(System.in);
-        n = scan.nextInt();
+        n = scan.nextDouble();
         for (int i = 1; i <= n; i++) {
             sum = sum + 1.0 / i;
         }
